flac: fix golangci lint issues
	frame/utf8.go:57:6: `decodeUTF8Int` is unused (deadcode)
	func decodeUTF8Int(r io.Reader) (n uint64, err error) {
		  ^
	internal/utf8/encode.go:32:16: unnecessary conversion (unconvert)
			bits = uint64(t2 | (x>>6)&mask2)
							 ^
	internal/utf8/encode.go:37:16: unnecessary conversion (unconvert)
			bits = uint64(t3 | (x>>(6*2))&mask3)
							 ^
	internal/utf8/encode.go:42:16: unnecessary conversion (unconvert)
			bits = uint64(t4 | (x>>(6*3))&mask4)
							 ^

+++ internal/utf8/encode.go
@@ -29,27 +29,27 @@ func Encode(w io.Writer, x uint64) error {
 		// if c0 == 110xxxxx
 		// total: 11 bits (5 + 6)
 		l = 1
-		bits = uint64(t2 | (x>>6)&mask2)
+		bits = t2 | (x>>6)&mask2
 	case x <= rune3Max:
 		// if c0 == 1110xxxx
 		// total: 16 bits (4 + 6 + 6)
 		l = 2
-		bits = uint64(t3 | (x>>(6*2))&mask3)
+		bits = t3 | (x>>(6*2))&mask3
 	case x <= rune4Max:
 		// if c0 == 11110xxx
 		// total: 21 bits (3 + 6 + 6 + 6)
 		l = 3
-		bits = uint64(t4 | (x>>(6*3))&mask4)
+		bits = t4 | (x>>(6*3))&mask4
 	case x <= rune5Max:
 		// if c0 == 111110xx
 		// total: 26 bits (2 + 6 + 6 + 6 + 6)
 		l = 4
-		bits = uint64(t5 | (x>>(6*4))&mask5)
+		bits = t5 | (x>>(6*4))&mask5
 	case x <= rune6Max:
 		// if c0 == 1111110x
 		// total: 31 bits (1 + 6 + 6 + 6 + 6 + 6)
 		l = 5
-		bits = uint64(t6 | (x>>(6*5))&mask6)
+		bits = t6 | (x>>(6*5))&mask6
 	case x <= rune7Max:
 		// if c0 == 11111110
 		// total: 36 bits (0 + 6 + 6 + 6 + 6 + 6 + 6)
